Add excluded_food tag for food excluded from the restaurant by default (#7808)
Adds excluded_food tag. Foods in this list will be disabled by default in newly-constructed restaurants, but still appear in the list and can be manually enabled if desired. Existing and upgraded restaurants are unaffected.
Sets most of the unhealthy vanilla foods and the expensive Minecolonies breads to excluded. (Golden and Chorus bread are out; Milky and Sweet bread are in. Golden apples are still in, but enchanted golden apples are out.)
Provides an easy mechanism for other huts to have custom defaults different from "clear everything".
Changes the creative paste to trigger an onUpgradeComplete 0 -> N rather than N -> N (was needed by an earlier version of this code; not actually needed by the final version but seemed like a good idea anyway so left it in).
Noticed but not investigated: onUpgradeComplete gets called twice for creative paste.

(cherry picked from commit fdf581ec648c5143170f0e980a780067b192a898)