Improve build tool detection of nearby buildings
Fixes the problem where the preview of a large building may not be visible until the build tool's active structure is already intersecting or is very close to the large building. This fix makes placing two large buildings (e.g.: a town hall and university) near each other a lot easier.