Tags and Recipes datagen (#8079)
Moves tags to datagen
Moves non-crafter recipes to datagen
Makes tags use standard loading mechanism to reduce complexity
This does reintroduce some glitches in the tag contents when used on a server and enabling/disabling a datapack while live, but a /reload fixes it (may require a double reload on edit). And presumably other mods suffer similar fates, so this shouldn't be too surprising. Works as expected in SP and when first connecting to an MP server, as far as I can tell (but not extensively tested)
Recipes now have unlock conditions, so will work in the vanilla recipe book
Most recipes are unlocked by having a build tool and sometimes another item in the inventory. Some of these might be a bit too harsh (e.g. farm won't unlock until you craft and hold a wooden hoe, which you might never do if you've already advanced beyond that) so possibly could be wound back a bit. But I did want to try to avoid unlocking absolutely everything on getting the build tool. (And of course the unlocks only matter if you're playing with doLimitedCrafting enabled, which is not the default.)
Some small tweaks to the tag contents
Enables smoking and campfire cooking recipes for mcol foods (from dough) ... not too likely to be used, since the baker will usually do the full round using the smelting recipe, but seemed useful for consistency.