Netherminer (#7884)
First off, we have a new building/worker pair, in the Nether Miner and Nether Mine
Wooden themed Nether Mine buildings level 1-5
Skins and models for the NetherMine hut block, and the Nether Miner
Research to unlock the building
Tweaks all over the place to make things work.
Added transparency layer support to hut blocks
Added support for invisible workers, to suppress many 'normal' behaviors while invisible.
Added support to recipe fulfillment to return the list of generated items without populating the worker's inventory.
Built a rudimentary 'simulation' capability based off of items generated from loottables.
Loot table generates tokens with information about the encounter
a battle is simulated, with ~50% chance to hit on either side.
For the duration of the simulation, the worker wears the armor/weapon, so damage is taken, and reduced.
Rounds continue to the death
Loot table also generates loot
Blocks are 'mined' with the appropriate tool, applying damage to the tool, and converting the blocks to the blocks loottable.
Number of items returned affects how long the overall simulation lasts.