Fix Compost Barrel logic to not void excess when extracting finished compost
Also changed the method of dispensing finished compost so if a side is given, finished compost is spawned as an item entity moving in the direction of the indicated side.
If null side is given, fixed old logic is invoked inserting results directly into the player's inventory with overflow being spawned on top of the barrel with no velocity.