Fix Compost Barrel logic to not void excess when extracting compost (#8119)
Fix Compost Barrel logic to not void excess when extracting compost
Also changed the method of dispensing finished compost:
If a side is given (the clicked side of the barrel), finished compost is spawned as an item entity moving in the direction of the indicated side.
If null side is given, fixed old logic is invoked inserting results directly into the player's inventory with overflow being spawned on top of the barrel with no velocity.