Allow hiring of employed workers (#8123)
This feature will allow huts hiring for "normal" (WorkerBuildingModuleView) jobs to hire workers that are already employed, firing them from their current job. Special huts (quarry, warehouse, school hiring students) will not be able to do this.*
Workers with jobs are listed after unemployed workers and workers who would normally be able to be hired (i.e. from the library).
This feature is disabled by default, with a button in the hire window screen to allow it. This button is disabled for huts that can't use this feature.