Evict the ~~very old netherminer~~ skeleton out of the vault (#8209)
Evict the very old netherminer skeleton out of the vault of the wooden nether mine:
I modified the schematics in a test world, and I noticed too late that the game was set to easy mode, from an earlier test.
After I created the vault and before I scanned it in, a skeleton spawned inside of the vault. This pull request removes him again.