Fixed issues with planter AI and resource scroll (#8259)
Upon unknown causes the plantation reports there's no soil positions, cause of this is unknown but for protection there's a safeguard around the random call so the range is always 0..1
Planter did not properly check inventory during the planting cycle, this caused him to walk around aimlessly not planting anything.
Resource scroll properly displays the building name it's working on again!