Better request handling (#8260)
This PR introduces resolver suitability. Resolvers can return a metric of how suited they are for the task.
This could be distance-based but also based on how long the list is.
It goes through the list of resolvers and for all resovlers of the same type it compares suitability.
Removes private crafting resolver from cook, only assistant can craft