new visitor and slight rework of loading (#8296)
Adds new Visitor Tobi
Queries the name off-thread as its very slow. Also avoids calling that too often if the player is offline or similar.
Properly use a player type model to display arms and legs correctly (and not just a mirror texture).
Jump of mount on unstuck
Fix crash with collission