adjust code, reduce cost (#8372)
The idea of this PR is to avoid an expensive check that has to run constantly for each citizen.
So, we do the check when the worker is first assigned the AI and then just make sure the building remains the same.

Aside of that, I slightly reduced the xp gathering. 4 blocks in each direction as the upper limit is enough.