Various fixes (#8381)
Prevents the worker making noises while in the vault (including damaged noises when attacked, and idle voice lines).
Fixes the worker being stuck in the vault but no longer actually on the nether trip if the world is saved and reloaded mid-trip.
On departing, the worker will now walk to the portal and light it, instead of remotely lighting it from their hut.
On returning, the worker will now immediately shut off the portal (when configured) before walking back to their hut (instead of usually never turning it off at all).
Tweaks the timing of returning slightly so that the worker should appear directly at the portal instead of tweening from the vault to the portal.
Plays the portal teleport sound effect when the worker leaves and returns.
A shorter custom sound effect might be nicer; the standard one is surprisingly long.