Fixing Amazon Spear To be Thrown From Thrower Position (#8645)
Sets position of spear entity when created, to be from thrower position instead of default 0,0,0
Moves the code for defining spear direction to be as part of constructor, since it could be abstracted there along with spear position
Confirmed that the spear is thrown as expected, and can be picked up. Also spawned an amazon raid to confirm that their spears cannot be picked up, and their spears also throw as expected.