Highway to the Lumber Zone (#8692)
Forester ignores all persistent (player-placed) leaf blocks, regardless of where they are.
If the restricted zone is enabled, the forester will no longer care that trees are within building bounds and will cut them anyway.
This does mean it's up to the player to designate their zone far enough away from their actual log-based building materials, or at least not allow non-persistent leaves to touch them.
But it now supports forestry buildings that have built-in treefarms, or that have weird shapes and placeholders such that the bounding box is not a good fit to the actual building shape. looks at Cavern Megahall
Adds a new forester option "Break Leaves" (off by default).
When enabled, the forester will take a bit more time explicitly breaking all the leaves on the tree, not just the logs. This may be useful to increase sapling yield when a fast leaf decay mod is not in use.
Note that when the shears option is also enabled, they will use shears on the leaves, which mostly defeats the point. So you will probably want to turn that off if you enable this setting.