New schematic fixes (#8726)
Fortress fixes:
Changes to the kelp farm, hopefully fixing the flowing water issue
Add a fence gate to the level 1 swineherder's hut
Sandstone fixes:
Replace the (broken) DO fences in the restaurant with vanilla fences
A rescan of the nether mine, as from screenshots on discord it looked like the vault moved between a version in minecolonies and the version on the schematics server,
and I can't recall rescanning the nether miner lately. It doesn't look like anything major changed, though
Jungle fixes:
Remove the superfluous alt farm. That farm was only added by mistake/misunderstanding and is completely equal to the "regular" farm anyway