The goggles, they do nothing (#8782)
Refactors the neighbor-building rendering to be a little more generic
Also fixes a bunch of things that seemed like bugs, though I didn't specifically verify in prior testing. For example it looked like if the blueprint future was particularly slow to complete, it would keep queueing up infinite requests. And if any blueprint failed to load it would not render any subsequent.
Adds the Builder's Goggles, which use this new implementation
image
image
They're a head-slot item which provides no armor at all.
They're also very fragile, and can't be repaired -- the glass and mechanisms are quite delicate, so only put them on when you really need to, and stay out of trouble while wearing them.
They are fairly well enchantable though, so you might be able to improve their durability a bit that way.
While worn, the goggles will display the bounding box and blueprint preview for any nearby active work orders in the colony.
There's a client config setting to specify how close is "nearby".
This includes both buildings and decorations. It will also render (box only) removal orders, and where the miner is currently working.
If you're holding shift, it will also show a box around the anchor (if it has one), helping you to find it again if you lose track of it.
This also highlights (only) the anchor of placed but as yet unbuilt buildings, which can help find hut blocks lost underground.