Crafting tools (#8762)
Allows crafterrecipes to specify a tool which is required to perform the craft.
The worker will request the tool if they don't have one; it follows the usual rules of min/max tool levels.
They won't attempt to craft a different recipe while waiting for the tool (unless some other recipe becomes "first fulfillable" in the meantime) -- i.e. having the tool is not considered a fulfillable precondition, unlike tools in the recipe grid itself.
While crafting, the worker will hold the tool.
It will take durability damage (or not, due to research/enchants) as items are crafted.
I haven't tested this, but I suspect if the tool breaks during the craft they might end up over-crafting, since they stop the craft there, the already-crafted items are not necessarily picked up for a while, and the original request (and count) will be restarted once they get another tool. Not sure if there's a good way to fix this, but it should otherwise be mostly harmless.
The required tool (if any) is shown in JEI and listed in the hut recipe list.
Changes the forester's log stripping recipes to require an axe.
Changes the farmer's pumpkin carving recipe to require shears.
Adds recipes to deoxidize copper blocks (and cut copper, at level 2+) to the mechanic
Also gives them a few bonus recipe slots to offset the additional recipes
Only the plain copper blocks can be deoxidized -- this doesn't support the full range of items, such as doing stairs and slabs too.
Waxing blocks is not inherent but is teachable (stonemason) as before.
Currently nobody has or can learn recipes to scrape off wax (but adding those might cause some duplicate output ambiguity or crafting loops?).