Search for buildings (#8802)
Adds a building browser / search.
With a hut block in hand, right-click in air:
image
After about 5-10s (might be faster/slower depending on CPU, disk, and number of installed style packs) it loads:
image
This shows the style pack, path, size, and level of all blueprints containing that hut.
Main building blueprints are in black
Parent building/deco blueprints (i.e. those that have that hut as a child) are in green.
Invisible buildings are hidden (in survival) or shown in blue (in creative).
The search is carried out purely client-side.
The search is cached (per login session) so if you reopen the window (even for a different hut) it will load immediately, but after logout it will take time the first time you search again.
This does mean that (unlike before caching) it won't update with recently-scanned buildings, but that's probably no great loss.