Rework of caledonia (#8863)
Rework of caledonia (Thanks to Thaylar!):
The Caledonia governor looked around and decided some changes needed making.
The governor instructed the builder to rid the town of the trap doors the general citizenry loved to play with so much.
The Enchanter, working with the governor, decided it was time for a remodel his interior.
The Governor got tired of hearing the Doctor complain and built an entirely new hospital.
The general citizenry was rather upset by the smell, but were greatly pleased when the Gravedigger climbed out of the new catacombs that had been dug.
The Gravedigger apologized profusely and laid the citizens beloved departed to rest in proper graves.
Through the Governor's extensive network of spies...uh Friends. A very friendly network. The Governor was able to entice an Alchemist and Netherworker to join the colony.
The Governor also got tired of looking at the colony monument, and instructed the Mystic to move their table to more inconspicuous location.
And last, but certainly not least, the Governor used a little extra tax revenue to build the colony a new Town Hall, befitting the glory and majesty of the Caledonia colony!
Note for players when they update past the version where this get merged
The following buildings may have issues due to moved hut blocks, so they may need to be deconstructed and rebuilt

Enchanter
Hospital
Mystic Site
The following buildings may have issues because they got renamed, and may need to be deconstructed and rebuilt
All alternate buildings (many got renamed from <building>alt to alt<building>)
Some levelable decorations got renamed
The following buildings got different dimensions, and might cause issues around the building
The graveyard got a catacomb, so it got expanded downwards
Review please