Fix NBT assignment corruption (#8905)
In some cases it's possible that NBT data gets corrupted and causes citizens their jobs to no longer match the expected job in the building they are assigned to.
This fix will prevent any loading issues by unassigning the worker when an improper link between the citizen it's job and building is found, to prevent any crashes.