Rack fixes in 1.19.2 (#9003)
Fix all broken racks in existing schematics, so that they won't visually overlap with other blocks anymore
(or other racks, causing glitching racks), or having weird interfaces due to block states being incorrect.
The racks were fixed using the following logic:
The racks that render as 2 blocks first check whether they face an invisible rack before. If that invisible rack was faced incorrectly, that is fixed (that should make sure the interface of that part of the rack opens correcty)
After that, it uses the old NBT tag relNeighbor to check whether it can find the blockrackair. It changes the faces of both parts if needed
Then it looks for other invisible (blockrackair) racks to pair up with, possibly in another direction (in the tiny chance both the direction and the relNeighbor were incorrect)
Next, it will do the same three steps with any racks, to fix racks glitching (those other ones are converted to blockrackair)
The remaining ones that need to be fixed are loose invisible racks. If those happen to be next to each other, they are paired together (it looks like the scan tool tries to do something like that as well when converting blocks to racks)
The last ones are converted to single racks.
I checked the blueprints in game, and they look much better now. In extreme cases, the racks may still look a little odd, where there are two single ones next to each other instead of a double rack. Those were likely very messed up, and may need a manual fix later.
I couldn't find any overlapping or wrong rack anymore
I splitted the changes up in multiple commits, one per style folder, just because there are a lot of changes