Don't allow kids to be born into explicitly manual housing. (#9044)
Adds a new Locked mode to residence/tavern, that otherwise mostly behaves like Manual.
If a residence/tavern is set explicitly to Manual Locked, then kids won't be born into it, even if it has spare beds.
Kids will still be born if set to Default any other mode, regardless of whether the colony as a whole is set to Manual or Auto.
Review please (could port)

In particular, this allows you to deliberately leave some housing vacant, such as leaving a tavern empty once you have level 4/5 housing, or having a single-person house at the edge of your colony for the colonist with the funny smell... (without making them unhappy that they only have a level 1 house)