Fix ignoring of NBT on research costs for tools as well (#8855)
Fixes the research item cost checking so that it ignores NBT on items with damage bars as apparently intended.
This does increase the risk that it might steal your enchanted or otherwise NBT-carrying tools, but it does still check for 0/matching damage so it won't steal anything that isn't fully repaired, so most in-use tools should be safe.
This should fix an incompatibility with Apotheosis and Quark and similar mods that like adding extra NBT to otherwise ordinary tools.