Recipe templates (#9164)
Adds "custom recipe templates". These are similar in format to custom recipes but with an extra wrapper that specifies a tag and filter -- for each item in the tag matching the filter, it will attempt to substitute the resource id into the recipe template contents, and if all of these substitutions result in valid items, the recipe is considered valid and added.
"type" is always "recipe-template"
"tag" specifies the name of a tag to search (without leading #)
"filter" is either a single string or an array of strings. Only items containing at least one of these strings are included.
Alternatively, it can be an object with "include" and "exclude" (each either string or list as above). As you'd expect, any substring in include adds it unless it also matches any substring in exclude.
"recipe" is the actual custom recipe, with properties as normal except that anywhere that an item id is specified, you may either specify a fixed id as usual or you can use one of the following substitutions (amongst other fixed characters):
[NS] => the namespace of the tag item
[PATH] => the path of the tag item
[PATH:abc=def] => the path of the tag item, replacing abc with def
Removes the old datagen recipes for log stripping.
Adds new (non-datagen) recipe templates for log stripping.
Notably, these add support for mangrove, and should in theory automatically extend to modded logs, provided they're added to the minecraft:logs tag and they fit the standard naming conventions (including having a wood form).
If auditcraftingtags is enabled, it logs status on login/reload (to minecraft logs, not files) to help confirm that it's generating the recipes you're expecting.
Special exception: an "alternate-output" that fails to generate a real item will be ignored (and logged) instead of rejecting the entire recipe -- unless everything fails and the whole recipe then has no outputs, in which case it's still rejected.
In order to support that, it's now allowed (only when using alternate-output) to not specify a result and only have alternates.