Cavern tweaks and Colonial fixes in 1.19.2 (#9242)
Cavern:
Match the number of racks in the alternate warehouse with the default warehouse
Small correction in the "standard" geology megahall, and three new alternatives to it
New Storage Megahall! Includes a megahall warehouse and 8 courier's huts (10 didn't fit. The other two should be made using the default or alternate courier's huts)
Colonial:
Fix the level 5 bakery requiring gold bread, for which you need a level 5 bakery