fix 9402 (#9403)
Fixes raider death messages, more specifically the duplicate death.attack. prefix. This was caused by including the death.attack. prefix in the damage source's msgId field, which is not what should be done. Vanilla adds the death.attack. prefix either in the getter or directly in the death message generation; this PR adjusts the relevant code to do so as well.