adjust nether colony spawn (#9765)
This PR reworks how we do the nether colony worldgen. Instead of having it specific for nether colonies, any structure with "allow_cave": true, can spawn in a cave like system without problems.

This also improves the nether colony spawning. Instead of a weird calc that tries to find 65 free space above lava and spawns it at the top. It tries 10 random locations and tries to find 32 air space and spawns it at the ground.