Rework the quarrier to support other builder modes in the future (#9967)
Reimplements the quarrier's logic using a new "LayerBlueprintIterator", which wraps around an existing iterator (at the moment, the default one).
This simplifies the code quite a bit, as this new iterator marks itself finished when the Y-level has been iterated over. The skip predicates don't need to block lower Y-levels anymore, and it only needs to check whether the iterator was finished when calculating resources for the current state, instead of also checking Y-levels there
The structureStep method has been reimplemented, now using the base class logic instead of using a copy of it. Three methods have been overridden to get the difference in logic, most notably goToNextStage. It makes sure the quarrier will do the next stage at the correct layer, and restarts the stages when the quarrier was not at the bottom yet