Fix guards unable to request shields (#10227)
With the tool rework shields were returning an item level of 1 by default, and can actually be levellable based on their durability to cope with modded shields better. The knight AI however was attempting to request shields in the range of level [0, 0]. Made the effective range for shields requests between 0 and max int.