Add crops to JEI (#10238)
Fixes an error parsing certain 1.20+ loot tables (causing some items to be missing in JEI).
Changes the crop loot generation to use definitions from the crops themselves rather than duplicating values. The drop probabilities should be the same as before.
Removed rice from sugar cane and added it to seagrass and small dripleaf.
Added "river" to humid biomes so that seagrass is actually available. (It's also still a temperate biome too.)
Added cabbage appearing in dungeon loot chests.
Prevent dropping crops if using shears or a silk touch tool.
Adds all growable crops to the existing JEI farmer page (with some display caveats).
image
For the MineColonies crops, hovering the "Biomes" label will show the biomes where this crop can be grown (and found):
image
I did have this one per line originally, which was easier to read, but this quickly overflowed the screen with some crops; that would only get worse when adding modded biomes.
Adds a "Colony Crops" page that displays how to obtain the MineColonies crops:
image
This only displays the blocks that do have MineColonies drops, but it does currently display all drops from that block, including vanilla and from other mods. I did consider filtering the display but this seemed ok as is, at least in mostly-vanilla.
It also assumes that all source blocks do have an item form. I did originally make it render the block directly but that didn't look as good.
The hoe is not required for any of these drops, but it's shown anyway as a hint that it improves your chances. The chance percentages shown are assuming that you do use the hoe.
The biomes are not directly visible here, but you can click on any item to see it in the Farmer's tab.