Raid commands update with optional raider amount and position arguments (#11287)
Updated raid command to include new amount and position arguments
Removed raid-all command in favor of selecting /mc colony raid all
Renamed @here and @mine to here and mine respectively, to match the naming of all
Made a MultiColonyIdArgument and updated the ColonyIdArgument to have identical logic when it comes to choosing colonies. The only addition that MultiColonyIdArgument has is the addition of the option to choose all
Created a base class for arguments that allows easier switching over several types of different values, you can provide multiple AllowedOption instances which match against the inputted argument, and see if the argument somehow matches the provided input. The argument then finds the first AllowedOption that matches the argument value, and stores this one as the argument result.
Removed extraneous overloads of the raiderEvent and setRaidTonight by using a container class to pass different settings

# Conflicts:
#	src/main/java/com/minecolonies/core/colony/events/raid/RaidManager.java
#	src/main/java/com/minecolonies/core/commands/arguments/ColonyIdArgument.java
#	src/main/java/com/minecolonies/core/commands/arguments/ModArgumentTypes.java
#	src/main/java/com/minecolonies/core/commands/citizencommands/CommandCitizenModify.java
#	src/main/java/com/minecolonies/core/commands/colonycommands/CommandRaid.java
#	src/main/java/com/minecolonies/core/commands/generalcommands/CommandRaidAll.java
#	src/main/resources/META-INF/accesstransformer.cfg